requirements.txt
Python==3.11.2
requests==2.28.2
python-dotenv==0.21.1